02-08-23

+++ src/flprogUart.h
@@ -20,8 +20,8 @@ public:
     void setPortParity(uint8_t parity);
     virtual void setDeviceName(String name);
     virtual void changePins(int16_t newRxPin, int16_t newTxPin);
-    virtual void changePort(uint16_t newPort);
-    virtual void changePinsAndPort(int16_t newPort, int16_t newRxPin, int16_t newTxPin);
+    virtual void changePort(uint8_t newPort);
+    virtual void changePinsAndPort(uint8_t newPort, int16_t newRxPin, int16_t newTxPin);
 
 protected:
     virtual Stream *stream() { return uartPort(); };
